feat: Break parsers out into separate files
