+++ src/parsers/all.js
@@ -1,0 +1,5 @@
+import CucumberParser from './cucumber'
+import ShardedParser from './sharded'
+import StandardParser from './standard'
+
+export default [CucumberParser, ShardedParser, StandardParser]
